+++ anyblok/bloks/anyblok_core/core/sqlbase.py
@@ -12,7 +12,6 @@ from anyblok.field import FieldException
 from ..exceptions import SqlBaseException
 from sqlalchemy.orm import aliased, ColumnProperty
 from sqlalchemy.sql.expression import true
-from sqlalchemy.sql import functions
 from sqlalchemy import or_, and_
 
 
@@ -161,40 +160,8 @@ class SqlMixin:
     def _fields_description(cls):
         """ Return the information of the Field, Column, RelationShip """
         Field = cls.registry.System.Field
-        Column = cls.registry.System.Column
-        RelationShip = cls.registry.System.RelationShip
-
-        def get_query(Model):
-            columns = [
-                Model.name.label('id'),
-                Model.label,
-                Model.ftype.label('type'),
-            ]
-            if Model is Column:
-                columns.append(Model.nullable)
-                columns.append(Model.primary_key)
-                columns.append(Model.remote_model.label('model'))
-            elif Model is RelationShip:
-                columns.append(Model.nullable)
-                columns.append(functions.literal_column('false').label(
-                    'primary_key'))
-                columns.append(Model.remote_model.label('model'))
-            elif Model is Field:
-                columns.append(
-                    functions.literal_column('true as nullable'))
-                columns.append(functions.literal_column('false').label(
-                    'primary_key'))
-                columns.append(functions.literal_column('null').label(
-                    'model'))
-
-            return Model.query(*columns).filter(
-                Model.model == cls.__registry_name__)
-
-        query = get_query(RelationShip).union_all(get_query(Column)).union_all(
-            get_query(Field))
-        fields2get = [x['name'] for x in query.column_descriptions]
-        return {x.id: {y: getattr(x, y) for y in fields2get}
-                for x in query.all()}
+        query = Field.query().filter(Field.model == cls.__registry_name__)
+        return {x.name: x._description() for x in query.all()}
 
     @classmethod
     def fields_description(cls, fields=None):